Not working yet but closer
